Change abstraction points for reverse engineering so SQL Server and SQlite don't both implement the same IMetadataReader => IModel call
